Convert styling to moduler css and also create a delete account compnent
