+++ frontend/src/pages/Home/setting/setting.module.css
@@ -482,150 +482,7 @@
 
 /*! Responsive Styles */
 
-@media screen and (max-width: 1247px) {
-    .homeSection__currentUserContainer {
-        display: none;
-    }
-
-    .homeSection__ShowPostContainer {
-        width: 100%;
-    }
-
-    .profilePostContainer__PostBox {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        justify-content: center;
-    }
-}
-
-@media screen and (max-width: 990px) {
-    .authContainer__LeftSideContainer {
-        display: none;
-    }
-
-    .__outlet {
-        display: inline-block;
-        width: calc(100% - 80px);
-        height: 100vh;
-    }
-
-    .navbar,
-    .nav {
-        width: 80px;
-    }
-
-    .nav {
-        align-items: center;
-        justify-content: center;
-    }
-
-    .navLink {
-        width: 80%;
-    }
-
-    .instaLogo {
-        display: none;
-    }
-
-    .__instaIcon_Hide {
-        display: block;
-        font-size: 5px;
-        margin-left: 1.5em;
-    }
-
-    .__navTitle {
-        display: none;
-    }
-
-    .__popupContainer {
-        bottom: 30px;
-        right: -260px;
-    }
-}
-
-@media screen and (max-width: 815px) {
-
-    .dashboard__currentUser__PostsContainer,
-    .dashboard__currentUser__infoContainer {
-        width: 100%;
-    }
-}
-
-@media screen and (max-width: 768px) {
-    .dashboard__currentUser__PostsContainer {
-        margin: 0px auto;
-    }
-
-    .infoContainer__userProfile {
-        width: 120px;
-        height: 120px;
-    }
-
-    .userBox__userActivityState {
-        border-top: 1px solid rgba(128, 128, 128, 0.671);
-        height: 80px;
-        position: absolute;
-        bottom: 0;
-        left: 0;
-        width: 100%;
-        margin: 0;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        padding: 0em 1em;
-    }
-
-    .userBox__activity {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        height: 100%;
-    }
-
-    .infoContainer__userBox {
-        width: calc(100% - 140px);
-        align-self: flex-start;
-    }
-
-    .userBox__userName {
-        padding: 0.6em 0em;
-    }
-
-    .userBox__fullName {
-        margin: 10px 0px;
-        font-size: 20px;
-    }
-
-    .postsContainer__navItemTExt {
-        display: none;
-    }
-
-    .dashboard__postsContainer_navbar {
-        width: 100%;
-    }
-
-    .PostsContainer__navItem {
-        width: 120px;
-    }
-
-    .postsContainer__navitemICON {
-        width: 28px;
-        padding: 0;
-        margin: 0;
-    }
-
-    .noPost__secondaryTest {
-        width: 95%;
-    }
-
-}
-
 @media screen and (max-width: 681px) {
-    .HomeSection__homePostCard {
-        width: 98%;
-    }
-
     .__EditProfile_form {
         width: 100%;
         padding: 1em 0em;
@@ -637,69 +494,6 @@
 }
 
 @media screen and (max-width: 641px) {
-    .dashboard__currentUser__infoContainer {
-        min-height: 375px;
-    }
-
-    .userBox__fullName {
-        width: 100%;
-        padding-left: 1em;
-        position: absolute;
-        top: 40%;
-        left: 0;
-        font-size: 24px;
-        font-weight: 500;
-        padding: 0 1em;
-    }
-
-    .userBox__userBIO {
-        margin: 0;
-        width: 100%;
-        position: absolute;
-        top: 55%;
-        left: 0;
-        font-size: 20px;
-        font-weight: 400;
-        padding: 0 1.2em;
-        word-wrap: break-word;
-    }
-
-    .infoContainer__userBox {
-        min-height: fit-content;
-    }
-
-    .userBox__userName {
-        font-size: 28px;
-        padding-left: 0.1em;
-    }
-
-    .userBox__editProfileButton {
-        margin-top: 20px;
-        top: 100%;
-        left: 0;
-    }
-
-    .CreatePost__box {
-        width: 95%;
-    }
-
-    .createPost__form {
-        width: 100%;
-        box-shadow: none;
-    }
-
-    .createPost__currentUserInfo {
-        padding: 0em;
-    }
-
-    .createPost__captionBox {
-        width: 100%;
-    }
-
-    .createPost__selectImageICON {
-        margin-left: 0em;
-    }
-
     .__settingSection__nav {
         padding: 0em 0.5em;
         gap: 10px;
@@ -711,146 +505,14 @@
     }
 }
 
-@media screen and (max-width: 550px) {
-
-    .homeContainer {
-        flex-direction: column;
-    }
-
-    .__outlet {
-        width: 100%;
-        height: calc(100vh - 130px);
-    }
-
-    /* ! This block holds the Header Component Design */
-    .__nav_Header {
-        position: sticky;
-        top: 0;
-        left: 0;
-        border-bottom: 1px solid rgb(170, 170, 170);
-        width: 100%;
-        height: 4em;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        padding: 0 1em;
-        z-index: 2;
-    }
-
-    .__nav_Logo {
-        width: 9em;
-    }
-
-    .navLink {
-        padding: 0;
-        justify-content: center;
-    }
-
-    .navbar,
-    .primary__navbar {
-        border-top: 1px solid rgb(179, 179, 179);
-        flex-direction: row;
-        width: 100%;
-        height: 4em;
-        position: absolute;
-        bottom: 0;
-    }
-
-    .__instaIcon_Hide,
-    .__navbar_moreButton,
-    .__popupContainer {
-        display: none;
-    }
-
-    .__SecondaryPopupContainer {
-        display: block;
-        top: 110%;
-        right: 10px;
-    }
-
-    .__secondaryMoreButton {
-        padding: 0;
-        margin: 0;
-        width: 50px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .profilePostContainer__PostBox {
-        grid-template-columns: repeat(2, 0.8fr);
-    }
-
-    .profilePostContainer__postCard {
-        height: 275px;
-    }
-
-    .otpVerifier__Form {
-        width: 95vw;
-    }
-
-    .__lazyLoader_title {
-        font-size: 28px;
-        letter-spacing: normal;
-    }
-}
-
 @media screen and (max-width: 500px) {
-    .Auth__UserLoginFormContainer {
-        width: 100vw;
-    }
-
+   
     .__DeleteAccoutPopup {
         width: 95%;
     }
 
     .__DeleteAccountPopup__LIST {
         width: 100%;
-    }
-
-    .authFormn_Box,
-    .authGotoSignUP_container,
-    .Get_App_Container {
-        width: 95%;
-        margin-left: auto;
-        margin-right: auto;
-    }
-
-    .infoContainer__userProfile {
-        width: 100px;
-        height: 100px;
-    }
-
-    .userBox__userName {
-        font-size: 20px;
-    }
-
-    .infoContainer__userBox {
-        width: calc(100% - 110px);
-    }
-
-    .userBox__editProfileButton {
-        width: 140px;
-    }
-
-    .userBox__fullName {
-        top: 35%;
-    }
-
-    .userBox__userBIO {
-        width: 100%;
-        padding-left: 0.9em;
-        top: 50%;
-        font-size: 16px;
-        line-height: 25px;
-    }
-
-    .userBox__activity {
-        font-size: 16px;
-    }
-
-    .PostsContainer__navItem {
-        width: 80px;
     }
 
     .OTPForm__input {
@@ -882,9 +544,4 @@
         width: 40px;
         height: 40px;
     }
-
-    .userBox__editProfileButton {
-        margin-left: 0;
-    }
-
 }
